src/core/pyproject.toml
[project]
name = "hpc_regression"
version = "0.1.0"
description = "Minimal runner harness using structlog"
requires-python = ">=3.10"
dependencies = [
  "structlog",
  "pyyaml",
]

[project.scripts]
hpc-runner = "hpc_regression.cli:main"
